:wrench: Add .gitignore for Git, Terraform, Helm, IntelliJ
Added a standard .gitignore file for Git, Terraform, Helm, and IntelliJ. This includes ignoring backup files created by Git, chart dependencies for Helm, user-specific and generated files for IntelliJ, and local directories and sensitive files for Terraform. The purpose of this .gitignore file is to avoid committing unnecessary or sensitive files that do not contribute to the source code, or that can cause conflicts between developers' environments.