:sparkles: Add AWS parameter store for Bokeh ECR URLs
This commit creates a Terraform file for AWS parameter stores. Specifically, it sets up two AWS SSM parameters to store the repository URLs for 'bokeh_user_ecr' and 'bokeh_post_ecr'. This is needed to ensure that these URLs are easily retrievable for future updates or usages in other parts of the application deployment process.