:recycle: Rename ECR resources to follow service naming convention
In this commit, the names associated with the Elastic Container Registry (ECR) resources and the corresponding Simple System Manager (SSM) parameters in the AWS Terraform scripts are updated. Specifically, "bokeh-user-ecr" and "bokeh-post-ecr" are renamed to "bokeh-user-service" and "bokeh-post-service", respectively. This change is made to align with our standard of naming services and maintain consistency across our infrastructure.