:sparkles: Add AWS provider to ECR repositories
The AWS provider was added to the Bokeh-user and Bokeh-post ECR repositories in the container_registry.tf file. This was necessary because it now aligns with the AWS provider declared in the providers.tf file in the aws region "us-east-1". This change is important for consistency in AWS provider usage across the Terraform configuration.