:bug: Update AWS SSM parameter value in Terraform script
Changed the value of "aws_ssm_parameter" "bokeh-post_service_ecr_url" in the Terraform script. It was previously reflecting the bokeh-user_service repository URI. It's now updated with the correct bokeh-post_service repository URI. This modification ensures that the proper AWS ECR public repository is referenced in the AWS SSM parameter store.