:wrench: Add environment variable for Spring profile in User-Service
This commit introduces a new environment variable `SPRING_PROFILES_ACTIVE` in the deployment configuration of the `user-service`. This variable is parameterized with values from `charts/values.yaml`. It allows the service to switch between different configurations based on the active Spring profile, improving flexibility in application settings management across different environments. The default profile value is set to 'prod'.