:wrench: Update user-service and add database environment variables
This commit updates the port number for user-service in the values.yaml file from 8000 to 8080. Additionally, it adds environment variables for the user's MySQL database to user-service-deployment.yaml. The environment variables include DB_HOST, DB_PORT, DB_USERNAME and DB_PASSWORD. These variable holds the values for the database host, port, username and password. This is needed for configuration and providing the appropriate credentials for the user-service to connect to the MySQL database effectively.