:memo: Update README for clarity and accuracy
The changes in the README.md file are primarily to improve the description of the project and update the dependencies. The previous sections 'Infra Architecture' and 'Dependency' were overly complex and contained outdated information. Now, the description better represents the project as a GitOps repository for Bokeh. Additionally, the updated dependencies (Kubernetes, Helm, Terraform) accurately reflect the tools used in the project.