:wrench: Add Redis support to user-service deployment
User-service now uses Redis as a caching layer for improved performance. Two new template files: `user-redis-svc.yaml` and `user-redis-statefulset.yaml` have been introduced to provision a new Redis service and its statefulset. The configurations for the Redis services are added within the `values.yaml` file. Moreover, environment variables for Redis host and port are added to the `user-service-deployment.yaml` file for the user-service to connect to